modified DROP_REST logic
Changed removed the NOT operator on validateRest(). Renamed the method
to onPerson().  I lowered the threshold so that it would transition
through drops from simply shaking. Continuously shaking with previous
logic would cause the app to transition to DROP_NOTIFIED. After change,
the app worked as intended, would transition through drops, but return
to MONITORING